Split session-repository into focused modules
Split the 1,754-line session-repository.ts into 4 focused files:

- session-repository.ts: Core CRUD + parse functions
- session-match-repository.ts: Match-related functions
- session-queue-repository.ts: Queue management functions
- session-summary-repository.ts: Session summary

Key improvements:
- Fixed N+1 query in deleteLastMatch by combining 4 separate lookups into 1 query with join
- Fixed N+1 query in recalcQueuePositions by replacing per-player UPDATE loop with batch UPDATE using SQL CASE
- Extracted shared query logic in getActiveSessionFull and getSessionById into getSessionPlayersWithDetails helper

Updated all imports to use the new barrel export.
