+++ apps/worker/src/repositories/index.ts
@@ -4,4 +4,4 @@ export * as matchRepository from "./match-repository";
 export * as seasonPlayerRepository from "./season-player-repository";
 export * as teamRepository from "./team-repository";
 export * as achievementRepository from "./achievement-repository";
-export * as sessionRepository from "./session-repository";
+export * as sessionRepository from "./session";
